feature/SKOOP-123 (#70)
* Endpoint to delete notifications.
* CommunityUserRegistration is removed transitively when AcceptanceToCommunityNotification is deleted.

+++ src/main/java/com/tsmms/skoop/notification/command/NotificationCommandService.java
@@ -1,5 +1,7 @@
 package com.tsmms.skoop.notification.command;
 
+import com.tsmms.skoop.communityuser.registration.AcceptanceToCommunityNotification;
+import com.tsmms.skoop.communityuser.registration.CommunityUserRegistrationRepository;
 import com.tsmms.skoop.notification.Notification;
 import com.tsmms.skoop.notification.NotificationRepository;
 import org.springframework.stereotype.Service;
@@ -13,9 +15,13 @@ import static java.util.Objects.requireNonNull;
 public class NotificationCommandService {
 
 	private final NotificationRepository notificationRepository;
+	// TODO Remove dependency on other domain repository
+	private final CommunityUserRegistrationRepository communityUserRegistrationRepository;
 
-	public NotificationCommandService(NotificationRepository notificationRepository) {
+	public NotificationCommandService(NotificationRepository notificationRepository,
+									  CommunityUserRegistrationRepository communityUserRegistrationRepository) {
 		this.notificationRepository = requireNonNull(notificationRepository);
+		this.communityUserRegistrationRepository = requireNonNull(communityUserRegistrationRepository);
 	}
 
 	@Transactional
@@ -25,11 +31,17 @@ public class NotificationCommandService {
 
 	@Transactional
 	public void delete(Notification notification) {
+		if (notification == null) {
+			throw new IllegalArgumentException("Notification cannot be null.");
+		}
+		if (notification instanceof AcceptanceToCommunityNotification) {
+			communityUserRegistrationRepository.delete(((AcceptanceToCommunityNotification) notification).getRegistration());
+		}
 		notificationRepository.delete(notification);
 	}
 
 	@Transactional
-	public void delete(Collection<Notification> notifications) {
+	public void deleteAll(Collection<Notification> notifications) {
 		notificationRepository.deleteAll(notifications);
 	}
